add background logic for @auth0/next-js (#1788)
* add background logic for @auth0/next-js

Because Next.JS has a server, the auth0 library uses a regular
authentication flow instead of the Single-Page Application setting which
was previously used. This allows us to encrypt sessions and use a local
routing system to interface with Auth0. It's also nice that Next.JS is a
first-class citizen in the Auth0 ecosystem.